requirements.txt
streamlit==1.31.0
numpy==1.26.3
python-dotenv==1.0.1
google-generativeai==0.3.2